gpt2zero2/pyproject.toml
[project]
name = "gpt2zero2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.6.0",
    "einops>=0.8.1",
    "equinox>=0.13.0",
    "flax>=0.10.6",
    "ipython>=9.4.0",
    "jax[cuda12]>=0.7.1",
    "matplotlib>=3.10.3",
    "memray>=1.17.2",
    "ml-collections>=1.1.0",
    "numpy>=2.2.6",
    "ruff>=0.11.12",
    "tiktoken>=0.9.0",
    "torchviz>=0.0.3",
    "tqdm>=4.67.1",
    "transformer-lens>=2.15.4",
    "transformers>=4.52.4",
    "wandb>=0.19.11",
]
